+++ test/test_handlers.py
@@ -117,24 +117,6 @@ def test_condition():
         handlers.condition(model, {'y': 3.})()
 
 
-def test_do():
-    def model():
-        s = numpyro.sample("s", dist.Normal(), constraint=constraints.positive)
-        z = numpyro.sample('z', dist.Normal(0., s))
-        return z ** 2
-
-    intervened_model = handlers.do(model, data={"z": 1.})
-    with handlers.trace() as exec_trace:
-        z_square = handlers.seed(intervened_model, 0)
-    assert exec_trace['z']['value'] != 1.
-    assert not exec_trace['z']['is_observed']
-    assert not exec_trace['z'].get('stop', None)
-    assert z_square == 1
-    # Raise ValueError when site is already observed.
-    with pytest.raises(ValueError):
-        handlers.do(model, data={'z': 3.})()
-
-
 def test_no_split_deterministic():
     def model():
         x = numpyro.sample('x', dist.Normal(0., 1.))
@@ -259,3 +241,63 @@ def test_plate_stack(shape):
 
     x = handlers.seed(guide, 0)()
     assert x.shape == shape
+
+
+@pytest.mark.parametrize('intervene,observe,flip', [
+    (True, False, False),
+    (False, True, False),
+    (True, True, False),
+    (True, True, True),
+])
+def test_counterfactual_query(intervene, observe, flip):
+    # x -> y -> z -> w
+
+    sites = ["x", "y", "z", "w"]
+    observations = {"x": 1., "y": None, "z": 1., "w": 1.}
+    interventions = {"x": None, "y": 0., "z": 2., "w": 1.}
+
+    def model():
+        with handlers.seed(rng_seed=0):
+            x = numpyro.sample("x", dist.Normal(0, 1))
+            y = numpyro.sample("y", dist.Normal(x, 1))
+            z = numpyro.sample("z", dist.Normal(y, 1))
+            w = numpyro.sample("w", dist.Normal(z, 1))
+            return dict(x=x, y=y, z=z, w=w)
+
+    if not flip:
+        if intervene:
+            model = handlers.do(model, data=interventions)
+        if observe:
+            model = handlers.condition(model, data=observations)
+    elif flip and intervene and observe:
+        model = handlers.do(
+            handlers.condition(model, data=observations),
+            data=interventions)
+
+    tr = handlers.trace(model).get_trace()
+    actual_values = tr.nodes["_RETURN"]["value"]
+    for name in sites:
+        # case 1: purely observational query like handlers.condition
+        if not intervene and observe:
+            if observations[name] is not None:
+                assert tr.nodes[name]['is_observed']
+                assert_allclose(observations[name], actual_values[name])
+                assert_allclose(observations[name], tr.nodes[name]['value'])
+            if interventions[name] != observations[name]:
+                assert_raises(AssertionError, assert_allclose, interventions[name], actual_values[name])
+        # case 2: purely interventional query like old handlers.do
+        elif intervene and not observe:
+            assert not tr.nodes[name]['is_observed']
+            if interventions[name] is not None:
+                assert_allclose(interventions[name], actual_values[name])
+            assert_raises(AssertionError, assert_allclose, observations[name], tr.nodes[name]['value'])
+            assert_raises(AssertionError, assert_allclose, interventions[name], tr.nodes[name]['value'])
+        # case 3: counterfactual query mixing intervention and observation
+        elif intervene and observe:
+            if observations[name] is not None:
+                assert tr.nodes[name]['is_observed']
+                assert_allclose(observations[name], tr.nodes[name]['value'])
+            if interventions[name] is not None:
+                assert_allclose(interventions[name], actual_values[name])
+            if interventions[name] != observations[name]:
+                assert_raises(AssertionError, assert_allclose, interventions[name], tr.nodes[name]['value'])
